fix: TypeScriptエラーを完全修正・型安全性の確保
- EntryLayoutコンポーネントのPromise/error型エラー修正
- EntryHeaderコンポーネントの型定義とProps整理
- 不要なportfolioコンポーネントの削除
- firebaseLoaderの型エクスポート整理
- index.ts/index.d.tsの型定義統一
- 全agritechコンポーネントの型安全性確保
