+++ src/components/agritech/EntryLayout.astro
@@ -27,7 +27,7 @@ const entryDescription =
   (autodescription ? smartTruncate(entry.body!, descriptionLenth) : null);
 ---
 
-<BaseLayout title={title} description={entryDescription} image={image?.src}>
+<BaseLayout title={title} description={entryDescription} image={image}>
   <section class="flex container p-4">
     <div class={`w-full ${actuallyHideToc ? "" : "md:col-9"}`}>
       <article class="">
@@ -123,17 +123,17 @@ const entryDescription =
               },
               body: JSON.stringify({ url })
             }),
-            new Promise((_, reject) => 
+            new Promise<never>((_, reject) => 
               setTimeout(() => reject(new Error('timeout')), 3000)
             )
           ]);
 
-          if (response.ok) {
-            const data = await response.json();
+          if ((response as Response).ok) {
+            const data = await (response as Response).json();
             return createLinkCardHTML(data.url, data.title || hostname, data.description || '', data.image || '', data.siteName || hostname);
           }
         } catch (error) {
-          console.log('Firebase Function not available, using offline fallback:', error.message);
+          console.log('Firebase Function not available, using offline fallback:', (error as Error).message);
         }
 
         // Offline fallback with enhanced site recognition
@@ -149,7 +149,7 @@ const entryDescription =
       return createLinkCardHTML(url, title, description, imageUrl, hostname);
     }
 
-    function getEnhancedTitle(urlObj) {
+    function getEnhancedTitle(urlObj: URL) {
       const hostname = urlObj.hostname;
       const pathname = urlObj.pathname;
       
@@ -178,7 +178,7 @@ const entryDescription =
       return hostname.replace('www.', '').split('.')[0];
     }
 
-    function getEnhancedDescription(urlObj) {
+    function getEnhancedDescription(urlObj: URL) {
       const hostname = urlObj.hostname;
       const pathname = urlObj.pathname;
       
@@ -198,7 +198,7 @@ const entryDescription =
       return `${hostname}のコンテンツ`;
     }
 
-    function getEnhancedImage(urlObj) {
+    function getEnhancedImage(urlObj: URL) {
       const hostname = urlObj.hostname;
       
       if (hostname.includes('github.com')) {
@@ -207,7 +207,7 @@ const entryDescription =
       return '';
     }
 
-    function createLinkCardHTML(url, title, description, imageUrl, siteName) {
+    function createLinkCardHTML(url: string, title: string, description: string, imageUrl: string, siteName: string) {
       const iconSvg = getIconSvg(siteName);
       const imageSection = imageUrl ? 
         `<div class="flex-shrink-0 w-20 h-20 bg-gray-100 dark:bg-gray-700 rounded overflow-hidden">
@@ -234,7 +234,7 @@ const entryDescription =
       `;
     }
 
-    function getIconSvg(hostname) {
+    function getIconSvg(hostname: string) {
       if (hostname.includes('github.com')) {
         return `<svg class="w-10 h-10 text-gray-700 dark:text-gray-300" fill="currentColor" viewBox="0 0 24 24"><path d="M12 0c-6.626 0-12 5.373-12 12 0 5.302 3.438 9.8 8.207 11.387.599.111.793-.261.793-.577v-2.234c-3.338.726-4.033-1.416-4.033-1.416-.546-1.387-1.333-1.756-1.333-1.756-1.089-.745.083-.729.083-.729 1.205.084 1.839 1.237 1.839 1.237 1.07 1.834 2.807 1.304 3.492.997.107-.775.418-1.305.762-1.604-2.665-.305-5.467-1.334-5.467-5.931 0-1.311.469-2.381 1.236-3.221-.124-.303-.535-1.524.117-3.176 0 0 1.008-.322 3.301 1.23.957-.266 1.983-.399 3.003-.404 1.02.005 2.047.138 3.006.404 2.291-1.552 3.297-1.23 3.297-1.23.653 1.653.242 2.874.118 3.176.77.84 1.235 1.911 1.235 3.221 0 4.609-2.807 5.624-5.479 5.921.43.372.823 1.102.823 2.222v3.293c0 .319.192.694.801.576 4.765-1.589 8.199-6.086 8.199-11.386 0-6.627-5.373-12-12-12z"/></svg>`;
       } else if (hostname.includes('youtube.com') || hostname.includes('youtu.be')) {
