+++ src/pages/agritech/[entry].astro
@@ -1,7 +1,8 @@
 ---
 import EntryLayout from "@components/agritech/EntryLayout.astro";
 import similerItems from "@lib/similarItems";
-import { getFirebaseBlogEntries, getFirebaseBlogEntry, type BlogEntry } from "@lib/firebaseLoader";
+import type { BlogEntry } from "@/types";
+import { getFirebaseBlogEntries, getFirebaseBlogEntry } from "@lib/firebaseLoader";
 
 export async function getStaticPaths() {
   const entries = await getFirebaseBlogEntries();
